requirements.txt
Flask~=2.0.1
Pillow~=8.3.1
numpy~=1.19.5
keras~=2.6.0
gunicorn
tensorflow-cpu
tensorflow_hub
